requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
pandas==1.4.3
pillow==10.2.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.1
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==2.2.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.5.0
typing_extensions==4.9.0
urllib3==2.2.1
wandb==0.17.0
kmeans1d==0.3.1
paho-mqtt==2.1.0
pyyaml==6.0.1
